Convert COOP+COEP tests to dispatcher.js
popup-test.js cannot be directly adapted for COOP+COEP tests. To solve
the issue we pass in a list of headers instead of simply the COOP
headers. This requires some rewriting but allows to convert these tests
to popup-test.js as well.

This enables the removal of several common.js functions in a followup:
https://chromium-review.googlesource.com/c/chromium/src/+/3679738

Bug: 1318373
Change-Id: I2dcb56c0e0080f0d3446fe66ce36a38d88380034